+++ app/components/Settings.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace app\components;
+
+use justcoded\yii2\settings\components\DbSettings;
+use justcoded\yii2\settings\forms\AppSettingsForm;
+
+/**
+ * Class Settings
+ *
+ * @property AppSettingsForm $app
+ *
+ * @package app\components
+ */
+class Settings extends DbSettings
+{
+
+}