+++ config/app-web.php
@@ -2,8 +2,8 @@
 
 \Yii::setAlias('@app', dirname(__DIR__) . '/app');
 
-$params = require __DIR__ . '/params.php';
 $db     = require __DIR__ . '/db.php';
+$params = require __DIR__ . '/params.php';
 
 dotenv()->required('APP_KEY')->notEmpty();
 
@@ -12,7 +12,7 @@ $config = [
 	'basePath'   => dirname(__DIR__) . '/app',
 	'runtimePath'   => dirname(__DIR__) . '/runtime',
 	'vendorPath'   => dirname(__DIR__) . '/vendor',
-	'bootstrap'  => ['log'],
+	'bootstrap'  => ['log', 'settings'],
 	'aliases'    => [
 		'@config'=> '@app/../config',
 		'@bower' => '@vendor/bower-asset',
@@ -78,8 +78,14 @@ $config = [
 				],
 			],
 		],
+		'settings' => [
+			'class'     => 'app\components\Settings',
+			'defaults' => $params,
+			'modelsMap' => [
+				'app' => 'justcoded\yii2\settings\forms\AppSettingsForm',
+			],
+		],
 	],
-	'params' => $params,
 ];
 
 if (YII_ENV_DEV) {